requirements.txt
dbt-core
dbt-bigquery
google-cloud-pubsub